==> Natural Language Processing/1-Data-Cleaning.ipynb <==
# Data Cleaning
## Introduction
This notebook goes through a necessary step of any data science project - data cleaning. Data cleaning is a time consuming and unenjoyable task, yet it's a very important one. Keep in mind, "garbage in, garbage out". Feeding dirty data into a model will give us results that are meaningless.

Specifically, we'll be walking through:

1. Getting the data - in this case, we'll be scraping data from a website
2. Cleaning the data - we will walk through popular text pre-processing techniques
3. Organizing the data - we will organize the cleaned data into a way that is easy to input into other algorithms

The output of this notebook will be clean, organized data in two standard text formats:

1. Corpus - a collection of text
2. Document-Term Matrix - word counts in matrix format

# Problem Statement
As a reminder, our goal is to look at transcripts of various comedians and note their similarities and differences. Specifically, I'd like to know if Ali Wong's comedy style is different than other comedians, since she's the comedian that got me interested in stand up comedy.

# Getting The Data
Luckily, there are wonderful people online that keep track of stand up routine transcripts. Scraps From The Loft makes them available for non-profit and educational purposes.

To decide which comedians to look into, I went on IMDB and looked specifically at comedy specials that were released in the past 5 years. To narrow it down further, I looked only at those with greater than a 7.5/10 rating and more than 2000 votes. If a comedian had multiple specials that fit those requirements, I would pick the most highly rated one. I ended up with a dozen comedy specials.

# required installation
!pip install bs4
!pip install html5lib
!pip install lxml

# necessary imports for web scrapping
import requests
from bs4 import BeautifulSoup
import pickle

# function to scrape data 
def url_to_transcript(url):
    page = requests.get(url).text
    soup = BeautifulSoup(page, "lxml")
    text = [p.text for p in soup.find_all('p')]
    print(url)
    return text

# URLs considered for generationg transcripts
urls = ['http://scrapsfromtheloft.com/2017/05/06/louis-ck-oh-my-god-full-transcript/',
        'http://scrapsfromtheloft.com/2017/04/11/dave-chappelle-age-spin-2017-full-transcript/',
        'http://scrapsfromtheloft.com/2018/03/15/ricky-gervais-humanity-transcript/',
        'http://scrapsfromtheloft.com/2017/08/07/bo-burnham-2013-full-transcript/',
        'http://scrapsfromtheloft.com/2017/05/24/bill-burr-im-sorry-feel-way-2014-full-transcript/',
        'http://scrapsfromtheloft.com/2017/04/21/jim-jefferies-bare-2014-full-transcript/',
        'http://scrapsfromtheloft.com/2017/08/02/john-mulaney-comeback-kid-2015-full-transcript/',
        'http://scrapsfromtheloft.com/2017/10/21/hasan-minhaj-homecoming-king-2017-full-transcript/',
        'http://scrapsfromtheloft.com/2017/09/19/ali-wong-baby-cobra-2016-full-transcript/',
        'http://scrapsfromtheloft.com/2017/08/03/anthony-jeselnik-thoughts-prayers-2015-full-transcript/',
        'http://scrapsfromtheloft.com/2018/03/03/mike-birbiglia-my-girlfriends-boyfriend-2013-full-transcript/',
        'http://scrapsfromtheloft.com/2017/08/19/joe-rogan-triggered-2016-full-transcript/']

# Comedian names
comedians = ['louis', 'dave', 'ricky', 'bo', 'bill', 'jim', 'john', 'hasan', 'ali', 'anthony', 'mike', 'joe']

# scraping data
transcripts = [url_to_transcript(u) for u in urls]

# Pickle files for later access
for i, c in enumerate(comedians):
    with open("transcripts/" + c + ".txt", "wb") as file:
        pickle.dump(transcripts[i], file)

# loading pickle files as dictionaries
data = {}
for i, c in enumerate(comedians):
    with open("transcripts/" + c + ".txt", "rb") as file:
        data[c] = pickle.load(file)

data = {}
for i, c in enumerate(comedians):
    data[c] = transcripts[i]

# checkin if data is loaded properly
data.keys()

data['louis'][0]

# Cleaning The Data
When dealing with numerical data, data cleaning often involves removing null values and duplicate data, dealing with outliers, etc. With text data, there are some common data cleaning techniques, which are also known as text pre-processing techniques.

With text data, this cleaning process can go on forever. There's always an exception to every cleaning step. So, we're going to follow the MVP (minimum viable product) approach - start simple and iterate. Here are a bunch of things you can do to clean your data. We're going to execute just the common cleaning steps here and the rest can be done at a later point to improve our results.

## Common data cleaning steps on all text:

* Make text all lower case
* Remove punctuation
* Remove numerical values
* Remove common non-sensical text (/n)
* Tokenize text
* Remove stop words

## More data cleaning steps after tokenization:

* Stemming / lemmatization
* Parts of speech tagging
* Create bi-grams or tri-grams
* Deal with typos
* And more...

# checking data 
print(next(iter(data.keys())))
print(next(iter(data.values())))

# Now our data is stored in a dictionary in the form key:comedian and value: transcript
# where the transcript is a list of all the paragraphs in the scraped file
# defining a function to convert the list of paragraph into a single string
def combine_text(text_list):
    combined_text = ' '.join(text_list)
    return combined_text

# applying the function to values of our dictionary
data_combined = {key: [combine_text(value)] for (key, value) in data.items()}

# now converting the dictionary into a pandas dataframe
import pandas as pd
pd.set_option('max_colwidth', 150)

df = pd.DataFrame.from_dict(data_combined).transpose()
df.columns = ['transcript']
df = df.sort_index()
df

#Applying round one of text cleaning
import re
import string

def clean_round_one(text):
    # converting whole text to lowercase
    text = text.lower()
    # removing all the parenthesis
    text = re.sub('\[.*?\]', '', text)
    # removing all the punctuation marks
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    # removing all the digits and alpha-numeric characters
    text = re.sub('\w*\d\w*', '', text)
    return text

round_one = lambda x : clean_round_one(x)

data_clean1 = pd.DataFrame(df.transcript.apply(round_one))
data_clean1

# Applying round two of text cleaning
def clean_round_two(text):
    # removing some additional punctuation and non-sensical text that was missed the first time around.
    text = re.sub('[‘’“”…]', '', text)
    # removing line breaks
    text = re.sub('\n', '', text)
    return text
    
round_two = lambda x: clean_round_two(x)

data_clean = pd.DataFrame(data_clean1.transcript.apply(round_two))
data_clean

# Organizing The Data
I have organized data in two standard text formats:

Corpus - a collection of text
Document-Term Matrix - word counts in matrix format

## Corpus
We already created a corpus in an earlier step. The definition of a corpus is a collection of texts, and they are all put together neatly in a pandas dataframe here.

# Let us add full names of comedians 
full_names = ['Ali Wong', 'Anthony Jeselnik', 'Bill Burr', 'Bo Burnham', 'Dave Chappelle', 'Hasan Minhaj',
              'Jim Jefferies', 'Joe Rogan', 'John Mulaney', 'Louis C.K.', 'Mike Birbiglia', 'Ricky Gervais']

df['full_name'] = full_names
df

# pickling the data for later use
with open("pickle files/" + "corpus.pkl", "wb") as file:
    pickle.dump(df, file)

## Document-Term Matrix
For better analysis, the text must be tokenized, meaning broken down into smaller pieces. The most common tokenization technique is to break down text into words. We can do this using scikit-learn's CountVectorizer, where every row will represent a different document and every column will represent a different word.

In addition, with CountVectorizer, we can remove stop words. Stop words are common words that add no additional meaning to text such as 'a', 'the', etc.

# creating document term matrix using CountVectorizer
from sklearn.feature_extraction.text import CountVectorizer

cv = CountVectorizer(stop_words = 'english')
model = cv.fit_transform(data_clean.transcript)
data_dtm = pd.DataFrame(model.toarray(), columns=cv.get_feature_names())
data_dtm.index = data_clean.index
data_dtm

# pickling the files
with open("pickle files/" + "cv.pkl", "wb") as file:
    pickle.dump(data_dtm, file)

with open("pickle files/" + "clean.pkl", "wb") as file:
    pickle.dump(data_clean, file)
