==> Natural Language Processing/5-Text-Generation.ipynb <==
# Text Generation
## Introduction
Markov chains can be used for very basic text generation. It thinks every word in a corpus as a state, making a simple assumption that the next word is only dependent on the previous word.

Although Markov chains don't generate text as well as deep learning.

## Select Text to Imitate
Currently, I would be generating text in the style of Ali Wong, hence extracting her transcripts.

# reading the corpus, including punctuation
import pandas as pd

data = pd.read_pickle('pickle files/corpus.pkl')
data

# extracting only Ali's script
ali_text = data.transcript.loc['ali']
ali_text[:200]

## Building a Markov Chain Function
We are going to build a simple Markov chain function that creates a dictionary:

* The keys should be all of the words in the corpus
* The values should be a list of the words that follow the keys

from collections import defaultdict

def markov_chain(text):
    # tokenize the text by word, though including punctuation marks
    words = text.split(' ')
    
    # initializing a default dictionary as required data structure
    my_dict = defaultdict(list)
    
    # creating a zipped list of all of the word pairs and putting them in word: words followed
    for curr_word, next_word in zip(words[0:-1], words[1:]):
        my_dict[curr_word].append(next_word)
        
    # converting the default dict into a dictionary
    my_dict = dict(my_dict)
    return my_dict

# applying markov_chain to ALi Wong's script
ali_dict = markov_chain(ali_text)
ali_dict

## Creating a  Text Generator
In the following piece of code, I am going to create a function that generates sentences. It will take two things as inputs:

* The dictionary you just created
* The number of words you want generated

import random

def generate_sentence(chain, count = 15):
    # Capitalize the first word
    word1 = random.choice(list(chain.keys()))
    sentence = word1.capitalize()
    
    # generating second word from the value list.
    # set the new word as first words and  repeat
    for i in range(count - 1):
        word2 = random.choice(chain[word1])
        word1 = word2
        sentence += ' ' + word2
        
    # ending with a full stop
    sentence += '.'
    return sentence

# generating similar text to Ali Wong
generate_sentence(ali_dict)
